Add JWT authorization and ownership checks for materials and upcycled…

+++ internal/upcycledproducts/service.go
@@ -1,5 +1,7 @@
 package upcycledproducts
 
+import "errors"
+
 func GetAllUpcycledProducts() []Upcycledproduct {
 	return FetchUpcyclerProducts()
 }
@@ -7,10 +9,43 @@ func GetAllUpcycledProducts() []Upcycledproduct {
 func CreateUpcycledProductsService(upcycledproduct *Upcycledproduct) error {
 	return SaveUpcycledProduct(upcycledproduct)
 }
-func UpdateUpcycledProductsService(id string, upcycledproduct *Upcycledproduct) error {
+func UpdateUpcycledProductsService(
+	id string,
+	userID uint,
+	upcycledproduct *Upcycledproduct,
+) error {
+
+	var existing Upcycledproduct
+
+	err := GetUpcycledProductByID(id, &existing)
+
+	if err != nil {
+		return err
+	}
+
+	if existing.UpcyclerID != userID {
+		return errors.New("you do not own this product")
+	}
+
 	return UpdateUpcycledProductRepository(id, upcycledproduct)
 }
 
-func DeleteUpcycledProductsService(id string) error {
+func DeleteUpcycledProductsService(
+	id string,
+	userID uint,
+) error {
+
+	var product Upcycledproduct
+
+	err := GetUpcycledProductByID(id, &product)
+
+	if err != nil {
+		return err
+	}
+
+	if product.UpcyclerID != userID {
+		return errors.New("you do not own this product")
+	}
+
 	return DeleteUpcycledProductsRepository(id)
 }
